requirements.txt
numpy~=1.20.1
pandas~=1.2.2
pymongo==3.10.1
python-dateutil~=2.7.3
requests~=2.25.1
tweepy==3.10.0
Flask~=1.1.2
python-dotenv~=0.19.2
discord~=1.7.3
Scrapy~=2.5.1
fpdf~=1.7.2